Separated OAIFetcher out
