+++ harvester/fetcher/__init__.py
@@ -13,16 +13,12 @@ import urlparse
 import urllib
 import tempfile
 import re
-from sickle import Sickle
-from sickle.models import Record as SickleDCRecord
-from sickle.utils import xml_to_dict
 import urllib
 import requests
 import logbook
 from logbook import FileHandler
 import solr
 import pysolr
-from urlparse import parse_qs
 from pymarc import MARCReader
 import pymarc
 import pynux.utils
@@ -34,6 +30,9 @@ import dplaingestion.couch
 from ..collection_registry_client import Collection
 from .. import config
 from deepharvest.deepharvest_nuxeo import DeepHarvestNuxeo
+from .fetcher import Fetcher
+from .fetcher import NoRecordsFetchedException
+from .oaifetcher import OAIFetcher
 
 EMAIL_RETURN_ADDRESS = os.environ.get('EMAIL_RETURN_ADDRESS',
                                       'example@example.com')
@@ -42,104 +41,24 @@ STRUCTMAP_S3_BUCKET = 'static.ucldc.cdlib.org/media_json'
 NUXEO_MEDIUM_IMAGE_URL_FORMAT = "https://nuxeo.cdlib.org/Nuxeo/nxpicsfile/default/{}/Medium:content/"
 NUXEO_S3_THUMB_URL_FORMAT = "https://s3.amazonaws.com/static.ucldc.cdlib.org/ucldc-nuxeo-thumb-media/{}"
 
-class NoRecordsFetchedException(Exception):
-    pass
-
-class Fetcher(object):
-    '''Base class for harvest objects.'''
-    def __init__(self, url_harvest, extra_data):
-        self.url = url_harvest
-        self.extra_data = extra_data
-        self.logger = logbook.Logger('FetcherBaseClass')
-
-    def __iter__(self):
-        return self
-
-    def next(self):
-        '''This returns a set of json objects.
-        Set can be only one json object.
-        '''
-        raise NotImplementedError
-
-def etree_to_dict(t):
-    d = {t.tag : map(etree_to_dict, t.iterchildren())}
-    d.update(('@' + k, v) for k, v in t.attrib.iteritems())
-    d['text'] = t.text
-    return d
-
-class SickleDIDLRecord(SickleDCRecord):
-    '''Extend the Sickle Record to handle oai didl xml.
-    Fills in data for the didl specific values
-
-    After Record's __init__ runs, the self.metadata contains keys for the
-    following DIDL data: DIDLInfo, Resource, Item, Component, Statement,
-    Descriptor
-    DIDLInfo contains created date for the data feed - drop
-    Statement wraps the dc metadata
-
-    Only the Resource & Component have unique data in them
-
-    '''
-    def __init__(self, record_element, strip_ns=True):
-        super(SickleDIDLRecord, self).__init__(record_element,
-                                                strip_ns=strip_ns)
-        #need to grab the didl components here
-        if not self.deleted:
-            didl = self.xml.find('.//{urn:mpeg:mpeg21:2002:02-DIDL-NS}DIDL')
-            didls = didl.findall('.//{urn:mpeg:mpeg21:2002:02-DIDL-NS}*')
-            for element in didls:
-                tag = re.sub(r'\{.*\}', '', element.tag)
-                self.metadata[tag] = etree_to_dict(element)
-
-
-class OAIFetcher(Fetcher):
-    '''Fetcher for oai'''
-    def __init__(self, url_harvest, extra_data):
-        super(OAIFetcher, self).__init__(url_harvest, extra_data)
-        # TODO: check extra_data?
-        self.oai_client = Sickle(self.url)
-        self._metadataPrefix = 'oai_dc'
-        # ensure not cached in module?
-        self.oai_client.class_mapping['ListRecords'] = SickleDCRecord
-        self.oai_client.class_mapping['GetRecord'] = SickleDCRecord
-        if extra_data: # extra data is set spec
-            if 'set' in extra_data:
-                params = parse_qs(extra_data)
-                self._set = params['set'][0]
-                self._metadataPrefix = params.get('metadataPrefix', ['oai_dc'])[0]
-            else:
-                self._set = extra_data
-            #if metadataPrefix=didl, use didlRecord for parsing
-            if self._metadataPrefix.lower() == 'didl':
-                self.oai_client.class_mapping['ListRecords'] = SickleDIDLRecord
-                self.oai_client.class_mapping['GetRecord'] = SickleDIDLRecord
-            self.records = self.oai_client.ListRecords(
-                                    metadataPrefix=self._metadataPrefix,
-                                                    set=self._set,
-                                                    ignore_deleted=True)
-        else:
-            self.records = self.oai_client.ListRecords(
-                                    metadataPrefix=self._metadataPrefix,
-                                                    ignore_deleted=True)
-
-
-    def next(self):
-        '''return a record iterator? then outside layer is a controller,
-        same for all. Records are dicts that include:
-        any metadata
-        campus list
-        repo list
-        collection name
-        '''
-        while True:
-            sickle_rec = self.records.next()
-            if not sickle_rec.deleted:
-                break #good record to harvest, don't do deleted
-                      # update process looks for deletions
-        rec = sickle_rec.metadata
-        rec['datestamp'] = sickle_rec.header.datestamp
-        rec['id'] = sickle_rec.header.identifier
-        return rec
+###class NoRecordsFetchedException(Exception):
+###    pass
+###
+###class Fetcher(object):
+###    '''Base class for harvest objects.'''
+###    def __init__(self, url_harvest, extra_data):
+###        self.url = url_harvest
+###        self.extra_data = extra_data
+###        self.logger = logbook.Logger('FetcherBaseClass')
+###
+###    def __iter__(self):
+###        return self
+###
+###    def next(self):
+###        '''This returns a set of json objects.
+###        Set can be only one json object.
+###        '''
+###        raise NotImplementedError
 
 
 class SolrFetcher(Fetcher):
@@ -1095,6 +1014,8 @@ def main(user_email, url_api_collection, log_handler=None, mail_handler=None,
     if my_mail_handler:
         my_mail_handler.pop_application()
     return ingest_doc_id, num_recs, harvester.dir_save, harvester
+
+__all__ = (Fetcher, NoRecordsFetchedException)
 
 if __name__ == '__main__':
     args = parse_args()
