+++ test/test_oai_fetcher.py
@@ -1,0 +1,61 @@
+# -*- coding: utf-8 -*-
+import shutil
+from unittest import TestCase
+from test.utils import ConfigFileOverrideMixin, LogOverrideMixin
+from test.utils import DIR_FIXTURES, TEST_COUCH_DASHBOARD, TEST_COUCH_DB
+from harvester.collection_registry_client import Collection
+import harvester.fetcher as fetcher
+import httpretty
+
+class HarvestOAIControllerTestCase(ConfigFileOverrideMixin, LogOverrideMixin, TestCase):
+    '''Test the function of an OAI fetcher'''
+    def setUp(self):
+        super(HarvestOAIControllerTestCase, self).setUp()
+
+    def tearDown(self):
+        super(HarvestOAIControllerTestCase, self).tearDown()
+        shutil.rmtree(self.controller.dir_save)
+
+    @httpretty.activate
+    def testOAIHarvest(self):
+        '''Test the function of the OAI harvest'''
+        httpretty.register_uri(httpretty.GET,
+                'http://registry.cdlib.org/api/v1/collection/',
+                body=open(DIR_FIXTURES+'/collection_api_test.json').read())
+        httpretty.register_uri(httpretty.GET,
+                'http://content.cdlib.org/oai',
+                body=open(DIR_FIXTURES+'/testOAC-url_next-0.xml').read())
+        self.collection = Collection('http://registry.cdlib.org/api/v1/collection/')
+        self.setUp_config(self.collection)
+        self.controller = fetcher.HarvestController('email@example.com', self.collection, config_file=self.config_file, profile_path=self.profile_path)
+        self.assertTrue(hasattr(self.controller, 'harvest'))
+        # TODO: fix why logbook.TestHandler not working for the previous logging
+        # self.assertEqual(len(self.test_log_handler.records), 2)
+        self.tearDown_config()
+
+
+
+# Copyright © 2016, Regents of the University of California
+# All rights reserved.
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+# - Redistributions of source code must retain the above copyright notice,
+#   this list of conditions and the following disclaimer.
+# - Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+# - Neither the name of the University of California nor the names of its
+#   contributors may be used to endorse or promote products derived from this
+#   software without specific prior written permission.
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
+# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+
